pyproject.toml
[project]
name = "sample-programs"
version = "2025.11.08"
description = "Sample Programs in Every Programming Language"
readme = ".github/README.md"
requires-python = ">=3.10,<4.0"
license = "MIT"

dependencies = [
    "ronbun (>=0.10.1,<0.11.0)",
    "glotter2 (>=0.13.0,<0.14.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

# Initiator of the collection
authors = [
    {name = "Jeremy Grifski", email = "jeremy.grifski@therenegadecoder.com"}
]

# Major contributors along the way
maintainers = [
    {name = "auroq"},
    {name = "rzuckerm"},
    {name = "stuin"},
    {name = "noah11012"},
    {name = "heksterb"},
    {name = "msj2"},
    {name = "chrboe"},
    {name = "2Clutch"},
    {name = "alope107"}
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
console_output_style = "count"
